requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
geetest==3.2.1
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
PyJWT==1.7.1
PyMySQL==0.9.3
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.25.3
